+++ Changelog.md
@@ -1,8 +1,14 @@
 # Change Log
 
+#### 0004 - V9 Compatibility Version
+
+- V9 Compatibility
+- Moon Lighting Cycles added
+- Max lighting level added
+
 #### 0003
 
-- Edge case bug causing update to fail resolves - requires Simple Calendar v1.3.66 or greater
+- Edge case bug causing update to fail resolved - requires Simple Calendar v1.3.66 or greater
 
 #### 0002